Implement next number algorithm for prepayments and assets

+++ src/main/java/io/github/erp/internal/utilities/NextIntegerFiller.java
@@ -23,17 +23,17 @@ import java.util.List;
 
 public class NextIntegerFiller {
 
-    public static int fillNext(List<Integer> numbers) {
+    public static long fillNext(List<Long> numbers) {
         if (numbers.isEmpty()) {
             return 0; // If the list is empty, start with 0
         }
 
         Collections.sort(numbers); // Sort the list of numbers
 
-        int nextNumber = numbers.get(0); // Start with the smallest number
+        long nextNumber = numbers.get(0); // Start with the smallest number
 
         for (int i = 1; i < numbers.size(); i++) {
-            int currentNumber = numbers.get(i);
+            long currentNumber = numbers.get(i);
             if (currentNumber - nextNumber > 1) {
                 return nextNumber + 1; // Found a gap, return the next number to fill
             }
@@ -45,11 +45,11 @@ public class NextIntegerFiller {
     }
 
     public static void main(String[] args) {
-        List<Integer> numbers = new ArrayList<>();
-        numbers.add(3);
-        numbers.add(6);
-        numbers.add(1);
-        numbers.add(4);
+        List<Long> numbers = new ArrayList<>();
+        numbers.add(3L);
+        numbers.add(6L);
+        numbers.add(1L);
+        numbers.add(4L);
         System.out.println("Next number to fill: " + fillNext(numbers)); // Output: 2
     }
 }
